04-09

+++ SCEA_API/src/main/java/br/unipar/scea_api/repositories/TipoPerguntaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoPerguntaRepository extends JpaRepository<TipoPergunta, Integer> {
+public interface TipoPerguntaRepository extends JpaRepository<TipoPergunta, Long> {
 }
